+++ src/components/formularios/index.ts
@@ -2,3 +2,4 @@ export { default as AfectadosForm } from "./AfectadosForm.vue";
 export { default as VinculadosForm } from "./VinculadosForm.vue";
 export { default as EfectosForm } from "./EfectosForm.vue";
 export { default as PersonalIntervinienteComponent } from "./PersonalIntervinienteComponent.vue";
+export { default as FechaUbicacionComponent } from "./FechaUbicacionComponent.vue";
